deps/v8/test/inspector/debugger/pause-expected.txt

Checks Debugger.pause

Running test: testPause
paused at:
#var a = 42;


Running test: testSkipFrameworks
paused at:
#var a = 239;


Running test: testSkipOtherContext1
paused at:
var a = #1;


Running test: testSkipOtherContext2
paused at:
#var a = 239;

paused at:
var a = #239;


Running test: testWithNativeBreakpoint
paused at:
#var a = 42;

paused at:
#var a = 42;

paused at:
#var a = 42;


Running test: testResumeAnotherGroup
paused at:
#var a = 42;

{
    error : {
        code : -32000
        message : Can only perform operation while paused.
    }
    id : <messageId>
}
{
    error : {
        code : -32000
        message : Can only perform operation while paused.
    }
    id : <messageId>
}
{
    error : {
        code : -32000
        message : Can only perform operation while paused.
    }
    id : <messageId>
}
{
    error : {
        code : -32000
        message : Can only perform operation while paused.
    }
    id : <messageId>
}

Running test: testDisableBreaksShouldCancelPause
paused at:
#debugger
